app/requirements.txt
Flask
mysql-connector-python
numpy
sklearn